RavenDB-8913: AllocatedSpaceInBytes has negative size in storage report

+++ src/Voron/Debugging/StorageReportGenerator.cs
@@ -124,7 +124,7 @@ namespace Voron.Debugging
             return journals.Select(journal => new JournalReport
             {
                 Number = journal.Number,
-                AllocatedSpaceInBytes = journal.JournalWriter.NumberOfAllocated4Kb * 4 * Constants.Size.Kilobyte
+                AllocatedSpaceInBytes = (long)journal.JournalWriter.NumberOfAllocated4Kb * 4 * Constants.Size.Kilobyte
             }).ToList();
         }
